requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.1
    # via requests
cmake==3.27.7
    # via triton
ffmpeg==1.4
    # via leapfrogai-backend-whisper (pyproject.toml)
filelock==3.13.1
    # via
    #   torch
    #   triton
grpcio==1.59.0
    # via
    #   grpcio-health-checking
    #   grpcio-reflection
    #   leapfrogai
grpcio-health-checking==1.59.0
    # via leapfrogai
grpcio-reflection==1.59.0
    # via leapfrogai
idna==3.4
    # via requests
jinja2==3.1.2
    # via torch
leapfrogai==0.3.3
    # via leapfrogai-backend-whisper (pyproject.toml)
lit==17.0.3
    # via triton
llvmlite==0.41.1
    # via numba
markupsafe==2.1.3
    # via jinja2
more-itertools==10.1.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numba==0.58.1
    # via openai-whisper
numpy==1.26.1
    # via
    #   numba
    #   openai-whisper
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
openai-whisper==20230918
    # via leapfrogai-backend-whisper (pyproject.toml)
protobuf==4.24.4
    # via
    #   grpcio-health-checking
    #   grpcio-reflection
    #   leapfrogai
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via tiktoken
sympy==1.12
    # via torch
tiktoken==0.3.3
    # via openai-whisper
torch==2.0.1
    # via
    #   openai-whisper
    #   triton
tqdm==4.66.1
    # via openai-whisper
triton==2.0.0
    # via
    #   openai-whisper
    #   torch
typing-extensions==4.8.0
    # via torch
urllib3==2.0.7
    # via requests
wheel==0.41.3
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11

# The following packages are considered to be unsafe in a requirements file:
# setuptools
